+++ native/lib/wasm.c
@@ -20,7 +20,7 @@
 #include <musl/math/frexp.c>
 #include <musl/math/modf.c>
 
-size_t strlen(const char* str) {
+unsigned long strlen(const char* str) {
     const char *s;
     for (s = str; *s; ++s);
     return (s - str);
